+++ packages/cli/src/zosfiles/compare/uss/UssFile.handler.ts
@@ -24,8 +24,9 @@ export default class UssFileHandler extends ZosFilesBaseHandler {
             stageName: TaskStage.IN_PROGRESS
         };
 
-        commandParameters.response.progress.startBar({ task });
 
+        // RETRIEVING USS FILE
+        commandParameters.response.progress.startBar({ task });
         const ussFileContentBuf1 = await Get.USSFile(session, commandParameters.arguments.ussFilePath1,
             {
                 binary: commandParameters.arguments.binary,
@@ -34,23 +35,20 @@ export default class UssFileHandler extends ZosFilesBaseHandler {
                 task: task
             }
         );
-
-
         commandParameters.response.progress.endBar();
-        commandParameters.response.progress.startBar({ task });
 
+
+        // RETRIEVING THE SECOND USS FILE TO COMPARE
+        task.statusMessage = "Retrieving second uss-file";
+        commandParameters.response.progress.startBar({ task });
         let binary2 = commandParameters.arguments.binary2;
         let encoding2 = commandParameters.arguments.encoding2;
-        const browserView = commandParameters.arguments.browserView;
-
         if (binary2 == undefined) {
             binary2 = commandParameters.arguments.binary;
         }
         if (encoding2 == undefined) {
             encoding2 = commandParameters.arguments.encoding;
         }
-
-        task.statusMessage = "Retrieving second uss-file";
         const ussFileContentBuf2 = await Get.USSFile(session, commandParameters.arguments.ussFilePath2,
             {
                 binary: binary2,
@@ -59,37 +57,30 @@ export default class UssFileHandler extends ZosFilesBaseHandler {
                 task: task
             }
         );
+        commandParameters.response.progress.endBar();
 
+
+        //CHECKING IF NEEDING TO SPLIT CONTENT STRINGS FOR SEQNUM OPTION
         let ussContentString1 = "";
         let ussContentString2 = "";
-
-        if(!commandParameters.arguments.seqnum){
-            const seqnumlen = 8;
-
-            const ussFileStringArray1 = ussFileContentBuf1.toString().split("\n");
-            for (const i in ussFileStringArray1) {
-                const sl = ussFileStringArray1[i].length;
-                const tempString = ussFileStringArray1[i].substring(0, sl - seqnumlen);
-                ussContentString1 += tempString + "\n";
-            }
-
-            const ussFileStringArray2 = ussFileContentBuf2.toString().split("\n");
-            for (const i in ussFileStringArray2) {
-                const sl = ussFileStringArray2[i].length;
-                const tempString = ussFileStringArray2[i].substring(0, sl - seqnumlen);
-                ussContentString2 += tempString + "\n";
-            }
+        const seqnumlen = 8;
+        if(commandParameters.arguments.seqnum === false){
+            ussContentString1 = ussFileContentBuf1.toString().split("\n")
+                .map((line)=>line.slice(0,-seqnumlen))
+                .join("\n");
+            ussContentString2 = ussFileContentBuf2.toString().split("\n")
+                .map((line)=>line.slice(0,-seqnumlen))
+                .join("\n");
         }
         else {
             ussContentString1 = ussFileContentBuf1.toString();
             ussContentString2 = ussFileContentBuf2.toString();
         }
 
-        //  CHECHKING IsetsF THE BROWSER VIEW IS TRUE, OPEN UP THE DIFFS IN BROWSER
-        if (browserView) {
 
+        // CHECKING IF THE BROWSER VIEW IS TRUE, OPEN UP THE DIFFS IN BROWSER
+        if (commandParameters.arguments.browserView) {
             await DiffUtils.openDiffInbrowser(ussContentString1, ussContentString2);
-
             return {
                 success: true,
                 commandResponse: "Launching uss files diffs in browser...",
@@ -97,15 +88,12 @@ export default class UssFileHandler extends ZosFilesBaseHandler {
             };
         }
 
-        let jsonDiff = "";
-        const contextLinesArg = commandParameters.arguments.contextlines;
 
-        jsonDiff = await DiffUtils.getDiffString(ussContentString1, ussContentString2, {
+        // RETURNING DIFF
+        const jsonDiff = await DiffUtils.getDiffString(ussContentString1, ussContentString2, {
             outputFormat: 'terminal',
-            contextLinesArg: contextLinesArg
+            contextLinesArg: commandParameters.arguments.contextLines
         });
-
-
         return {
             success: true,
             commandResponse: jsonDiff,